Add Travis-CI integration [WIP, comments below not all true]
This is heavily based on what I contributed to ostree, which in turn is
an improved version of what we use in dbus. .travis.yml is
Travis-specific, but the intention is that tests/ci-* could be reused
with other CI systems.

It runs a build and automated tests (make check, make distcheck,
make install, make installcheck, gnome-desktop-testing-runner)
on Ubuntu 14.04 'trusty' (old LTS and the newest thing natively
available on Travis-CI), Ubuntu 16.04 'xenial' (current LTS),
Debian 8 'jessie' (stable) and Debian unstable, which should give us
a reasonable spread of dependencies' versions. It should be reasonably
straightforward to add Docker containers for non-Debian-based
distributions if desired.

Because there are no official backports of libgsystem and ostree to
jessie, trusty or xenial, and Alex's PPA doesn't always track the
required version, I'm building a suitable version from source
if required.

Signed-off-by: Simon McVittie <smcv@debian.org>