run: Create a shared XDG_RUNTIME_DIR for each app-ID
Like $XDG_RUNTIME_DIR/app/$FLATPAK_ID, this is shared between all
instances of the app, except for subsandboxed instances created by
flatpak-spawn --sandbox or equivalent. Unlike
$XDG_RUNTIME_DIR/app/$FLATPAK_ID, it does not exist at an equivalent
path on the host and in the sandboxed app.

Note that I'm putting this in a subdirectory,
$XDG_RUNTIME_DIR/.flatpak/$FLATPAK_ID/xdg-run on the host, instead of
creating it directly in $XDG_RUNTIME_DIR/.flatpak/$FLATPAK_ID. This
allows $XDG_RUNTIME_DIR/.flatpak/$FLATPAK_ID/ to be used to contain
other per-app-ID things, including references to files on the host
system (such as a private directory for /tmp or /dev/shm) that would
not be safe if the sandboxed app was able to modify them.

Resolves: https://github.com/flatpak/flatpak/issues/4120
Signed-off-by: Simon McVittie <smcv@collabora.com>