run: Add a --clear-env option, to be used by the portal
This means the portal can implement FLATPAK_SPAWN_FLAGS_CLEAR_ENV
without accidentally removing an LD_LIBRARY_PATH that might be
necessary to run Flatpak itself.

Also, non-application processes such as the D-Bus proxy will now
inherit environment variables from `flatpak run` instead of having
a cleared environment, which is a return to the behaviour they would
have had before CVE-2021-21261 was fixed.

The LD_LIBRARY_PATH is still not passed to bwrap(1), even for a
non-setuid bwrap for which it will not be ignored and might be
necessary. This can be done when we depend on a version of bwrap that
has a --clearenv option.

Signed-off-by: Simon McVittie <smcv@collabora.com>